// src/types/prompt.ts
export type PromptOption = {
  description?: string;
  label: string;
  value: string;
};

export type FashionPromptInput = {
  gender: string;
  height: number;
  weight: number;
  age: number;
  skinTone: string;
  hairColor: string;
  hairStyle: string;
  eyeColor: string;
  bodyType: string;
  camera: string;
  lens: string;
  position: string;
  lighting: string;
  clothingDesc: string;
  extraDetails: string;
};